# backend/docs/swagger.yaml
basePath: /api
definitions:
  model.AddFriendReq:
    properties:
      receiver_id:
        example: "0"
        type: string
      sender_name:
        maxLength: 64
        type: string
      verification_message:
        maxLength: 128
        type: string
    required:
    - receiver_id
    - sender_name
    type: object
...
  model.FriendRequestListResp:
    properties:
      created_at:
        type: string
      request_id:
        example: "0"
        type: string
      sender_id:
        example: "0"
        type: string
      sender_name:
        type: string
      status:
        type: string
      verification_message:
        type: string
    type: object
  model.FriendshipListResp:
    properties:
      friend_id:
        example: "0"
        type: string
      friend_remark:
        type: string
      friendship_id:
        description: friend_ships 表的主键
        example: "0"
        type: string
    type: object
  model.IDReq:
    properties:
      id:
        example: "0"
        type: string
    required:
    - id
    type: object
...
  model.RemarkReq:
    properties:
      remark:
        maxLength: 64
        type: string
    required:
    - remark
    type: object
...
  model.SendTextReq:
    properties:
      content:
        maxLength: 1024
        type: string
      conversation_id:
        example: "0"
        type: string
    required:
    - content
    - conversation_id
    type: object
...
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is the API documentation for the Inquire backend.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Inquire API
  version: "1.0"
paths:
...
  /auth/files/{message_id}:
    get:
      consumes:
      - application/json
      description: Download a file by message ID
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 消息ID
        in: path
        name: message_id
        required: true
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: 文件内容
          schema:
            type: file
        "400":
          description: message_id错误
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: 非法文件路径
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: 文件不存在
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/response.Response'
      summary: 下载文件
      tags:
      - message
  /auth/friendship_requests:
    get:
      consumes:
      - application/json
      description: 获取当前收到的好友申请列表
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            items:
              $ref: '#/definitions/model.FriendRequestListResp'
            type: array
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/response.Response'
      summary: 获取好友申请列表
      tags:
      - friendship_request
    post:
      consumes:
      - application/json
      description: 发送好友申请给指定用户
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 添加好友请求体
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/model.AddFriendReq'
      produces:
      - application/json
      responses:
        "201":
          description: 发送成功
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: 发送失败
          schema:
            $ref: '#/definitions/response.Response'
        "409":
          description: 请勿重复发送
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/response.Response'
      summary: 发送好友申请
      tags:
      - friendship_request
  /auth/friendship_requests/{request_id}:
    delete:
      consumes:
      - application/json
      description: 删除或拒绝指定的好友申请
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 申请ID
        in: path
        name: request_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: 删除成功
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: requestID错误
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/response.Response'
      summary: 删除好友申请
      tags:
      - friendship_request
    post:
      consumes:
      - application/json
      description: 接受指定的好友申请
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 申请ID
        in: path
        name: request_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: 接受成功
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: requestID错误
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/response.Response'
      summary: 接受好友申请
      tags:
      - friendship_request
  /auth/friendships:
    get:
      consumes:
      - application/json
      description: 获取当前用户的好友列表
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            items:
              $ref: '#/definitions/model.FriendshipListResp'
            type: array
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/response.Response'
      summary: 获取好友列表
      tags:
      - friendship
  /auth/friendships/{friend_id}:
    delete:
      consumes:
      - application/json
      description: 根据好友ID删除好友
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 好友ID
        in: path
        name: friend_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: 删除成功
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: ID错误或好友不存在
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/response.Response'
      summary: 删除好友
      tags:
      - friendship
  /auth/friendships/remark/{friend_id}:
    post:
      consumes:
      - application/json
      description: 修改好友的备注信息
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 好友ID
        in: path
        name: friend_id
        required: true
        type: string
      - description: 修改备注请求体
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/model.RemarkReq'
      produces:
      - application/json
      responses:
        "200":
          description: 修改成功
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: ID错误或输入不合法
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/response.Response'
      summary: 修改好友备注
      tags:
      - friendship
...
  /auth/messages/delete:
    delete:
      consumes:
      - application/json
      description: Delete a message
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 消息ID
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/model.IDReq'
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: json解析出错
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/response.Response'
      summary: 删除消息
      tags:
      - message
  /auth/messages/file:
    post:
      consumes:
      - multipart/form-data
      description: Send a file to a conversation
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 会话ID
        in: formData
        name: conversation_id
        required: true
        type: string
      - description: 文件
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: 发送成功
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/response.Response'
      summary: 发送文件
      tags:
      - message
  /auth/messages/recall:
    delete:
      consumes:
      - application/json
      description: Recall a sent message
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 消息ID
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/model.IDReq'
      produces:
      - application/json
      responses:
        "201":
          description: 撤回成功
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: json解析出错
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/response.Response'
      summary: 撤回消息
      tags:
      - message
  /auth/messages/text:
    post:
      consumes:
      - application/json
      description: Send a text message to a conversation
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 发送文本消息请求体
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/model.SendTextReq'
      produces:
      - application/json
      responses:
        "201":
          description: 发送成功
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: json解析出错
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/response.Response'
      summary: 发送文本消息
      tags:
      - message
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"